04-api-gateway/production/requirements.txt
flask==3.0.3
pyjwt==2.8.0
python-dotenv==1.0.1
psutil==6.0.0
